+++ source/projects/py/tests/_old/py_call.c
@@ -1,0 +1,93 @@
+// replaced by pure python call function
+
+/**
+ * @brief Converts a Max list to call a python function with arguments
+ *
+ * @param x pointer to object structure
+ * @param s symbol
+ * @param argc atom argument count
+ * @param argv atom argument vector
+ * @return t_max_err error code
+ */
+t_max_err py_call(t_py* x, t_symbol* s, long argc, t_atom* argv)
+{
+    PyGILState_STATE gstate;
+    gstate = PyGILState_Ensure();
+
+    char* callable_name = NULL;
+    PyObject* py_argslist = NULL;
+    PyObject* pval = NULL;
+    PyObject* py_callable = NULL;
+    // python list
+    PyObject* py_args = NULL; // python tuple
+
+    // first atom in argv must be a symbol
+    if (argv->a_type != A_SYM) {
+        py_error(x, "first atom must be a symbol!");
+        goto error;
+
+    } else {
+        callable_name = atom_getsym(argv)->s_name;
+        py_log(x, "callable_name: %s", callable_name);
+    }
+
+    py_callable = PyRun_String(callable_name, Py_eval_input, x->p_globals,
+                               x->p_globals);
+    if (py_callable == NULL) {
+        py_error(x, "could not evaluate %s", callable_name);
+        goto error;
+    }
+
+    py_argslist = py_atoms_to_list(x, argc, argv, 1);
+    if (py_argslist == NULL) {
+        py_error(x, "atom to py list conversion failed");
+        goto error;
+    }
+
+    py_log(x, "length of argc:%ld list: %d", argc, PyList_Size(py_argslist));
+
+    // convert py_args to tuple
+    py_args = PyList_AsTuple(py_argslist);
+    if (py_args == NULL) {
+        py_error(x, "unable to convert args list to tuple");
+        goto error;
+    }
+
+    pval = PyObject_CallObject(py_callable, py_args);
+    if (!PyErr_ExceptionMatches(PyExc_TypeError)) {
+        if (pval == NULL) {
+            py_error(x, "unable to apply callable(*args)");
+            goto error;
+        }
+        goto handle_output;
+    }
+    PyErr_Clear();
+
+    pval = PyObject_CallFunctionObjArgs(py_callable, py_argslist, NULL);
+    if (pval == NULL) {
+        py_error(x, "could not retrieve result of callable(list)");
+        goto error;
+    }
+    goto handle_output; // this is redundant but safer in case code is added
+
+handle_output:
+    py_handle_output(x, pval);
+    // success cleanup
+    Py_XDECREF(py_callable);
+    Py_XDECREF(py_argslist);
+    PyGILState_Release(gstate);
+    py_bang_success(x);
+    return MAX_ERR_NONE;
+
+error:
+    py_handle_error(x, "anything %s", s->s_name);
+    // cleanup
+    Py_XDECREF(py_callable);
+    Py_XDECREF(py_argslist);
+    Py_XDECREF(pval);
+    PyGILState_Release(gstate);
+    py_bang_failure(x);
+    return MAX_ERR_GENERIC;
+}
+
+