+++ source/projects/zedit/web/editor.mjs
@@ -2,16 +2,18 @@ import { EditorView, basicSetup } from "codemirror";
 import { python } from "@codemirror/lang-python";
 import { oneDark } from "@codemirror/theme-one-dark";
 
-// import "xterm/css/xterm.css";
-import { Terminal } from "xterm";
-import { Readline } from "xterm-readline";
-import { FitAddon } from "xterm-addon-fit";
 
 let editor = new EditorView({
     extensions: [basicSetup, python(), oneDark],
     doc: "# python code here\n",
-    parent: document.body,
+    parent: document.getElementById("editor"),
 });
+
+function log_msg(msg) {
+    document.getElementById("msg").innerHTML = msg;
+    $("#msg").fadeIn(2000, "linear");
+    $("#msg").fadeOut(5000, "linear");
+}
 
 function openCode() {
     var input = document.createElement("input");
@@ -44,7 +46,7 @@ function openCode() {
 
     input.click();
 
-    document.getElementById("msg").innerHTML = "file opened";
+    log_msg("file opened");
 }
 var open_btn = document.getElementById("open_btn");
 open_btn.addEventListener("click", openCode);
@@ -65,7 +67,7 @@ function saveCode() {
     // .then((json) => console.log(json));
 
     console.log(editor.state.doc.toString());
-    document.getElementById("msg").innerHTML = "file saved";
+    log_msg("file saved");
 }
 var save_btn = document.getElementById("save_btn");
 save_btn.addEventListener("click", saveCode);
@@ -86,7 +88,7 @@ function runCode() {
     // .then((json) => console.log(json));
 
     console.log(editor.state.doc.toString());
-    document.getElementById("msg").innerHTML = "code run";
+    log_msg("code run");
 }
 var run_btn = document.getElementById("run_btn");
 run_btn.addEventListener("click", runCode);
@@ -94,67 +96,141 @@ run_btn.addEventListener("click", runCode);
 // ----------------------------------------------------------------------
 // terminal
 
-const term = new Terminal({
-    theme: {
-        background: "#191A19",
-        foreground: "#F5F2E7",
-    },
-    cursorBlink: true,
-    cursorStyle: "block",
+$(function () {
+    $("#terminal").terminal(
+        {
+            hello: function (what) {
+                this.echo("Hello, " + what + ". Wellcome to this terminal.");
+            },
+
+            cat: function (width, height) {
+                return $(
+                    '<img src="https://placekitten.com/' +
+                        width +
+                        "/" +
+                        height +
+                        '">'
+                );
+            },
+
+            title: function () {
+                return fetch("https://terminal.jcubic.pl")
+                    .then((r) => r.text())
+                    .then((html) => html.match(/<title>([^>]+)<\/title>/)[1]);
+            },
+
+            // opts like argument parsing (-a / --a)
+            demo: function (...args) {
+                const options = $.terminal.parse_options(args);
+                return options;
+            },
+
+            py: {
+                eval: function (arg) {},
+                exec: function (arg) {},
+                run: function (arg) {},
+                load: function (arg) {},
+                save: function (arg) {},
+            },
+
+            name: function (name) {
+                this.push(
+                    function (last_name) {
+                        if (last_name) {
+                            this.echo(
+                                "Your name is " + name + " " + last_name
+                            ).pop();
+                        }
+                    },
+                    {
+                        prompt: "last name: ",
+                    }
+                );
+            },
+        },
+        {
+            keymap: {
+                "CTRL-C": function (e, original) {
+                    this.echo("my shortcut");
+                },
+            },
+            checkArity: false,
+            completion: true,
+            greetings: "Python 3.11.3\n",
+        }
+    );
 });
+$.terminal.syntax("python");
+$.terminal.prism_formatters = {
+    prompt: true,
+    echo: true,
+    animation: true, // will be supported in version >= 2.32.0
+    command: true,
+};
 
-// term.write("Python 3.11.2 (web-editor)!\n");
+document.getElementById("default-tab").click();
 
-const fitAddon = new FitAddon();
-const readlineAddon = new Readline();
+// const term = new Terminal({
+//     theme: {
+//         background: "#191A19",
+//         foreground: "#F5F2E7",
+//     },
+//     cursorBlink: true,
+//     cursorStyle: "block",
+// });
 
-term.loadAddon(fitAddon);
-term.loadAddon(readlineAddon);
-term.open(document.getElementById("terminal"));
-fitAddon.fit();
-term.focus();
+// // term.write("Python 3.11.2 (web-editor)!\n");
 
-readlineAddon.setCheckHandler((text) => {
-    let trimmedText = text.trimEnd();
-    // if (trimmedText.endsWith("&&")) {
-    if (trimmedText.endsWith("\\")) {
-        return false;
-    }
-    return true;
-});
+// const fitAddon = new FitAddon();
+// const readlineAddon = new Readline();
 
-function readLine() {
-    readlineAddon.read(">>> ").then(sendReplEntry);
-}
+// term.loadAddon(fitAddon);
+// term.loadAddon(readlineAddon);
+// term.open(document.getElementById("terminal"));
+// fitAddon.fit();
+// term.focus();
 
-// function processLine(text) {
-// readlineAddon.println("");
-// readlineAddon.println("you entered: \n" + text.replace("\\\\", ""));
-// setTimeout(readLine);
+// readlineAddon.setCheckHandler((text) => {
+//     let trimmedText = text.trimEnd();
+//     // if (trimmedText.endsWith("&&")) {
+//     if (trimmedText.endsWith("\\")) {
+//         return false;
+//     }
+//     return true;
+// });
+
+// function readLine() {
+//     readlineAddon.read(">>> ").then(sendReplEntry);
 // }
 
-function respondReplEntry(json) {
-    readlineAddon.println(JSON.stringify(json));
-    setTimeout(readLine);
-}
+// // function processLine(text) {
+// // readlineAddon.println("");
+// // readlineAddon.println("you entered: \n" + text.replace("\\\\", ""));
+// // setTimeout(readLine);
+// // }
 
-function sendReplEntry(text) {
-    // readlineAddon.println("");
-    // readlineAddon.println("you entered: \n" + text.replace("\\\\", ""));
-    // setTimeout(readLine);
-    fetch("/api/repl/send", {
-        method: "POST",
-        body: JSON.stringify({
-            content: text,
-        }),
+// function respondReplEntry(json) {
+//     readlineAddon.println(JSON.stringify(json));
+//     setTimeout(readLine);
+// }
 
-        headers: {
-            "Content-type": "application/json; charset=UTF-8",
-        },
-    })
-        .then((response) => response.json())
-        // .then((json) => console.log(json));
-        .then((json) => respondReplEntry(json));
-}
+// function sendReplEntry(text) {
+//     // readlineAddon.println("");
+//     // readlineAddon.println("you entered: \n" + text.replace("\\\\", ""));
+//     // setTimeout(readLine);
+//     fetch("/api/repl/send", {
+//         method: "POST",
+//         body: JSON.stringify({
+//             content: text,
+//         }),
 
-readLine();
+//         headers: {
+//             "Content-type": "application/json; charset=UTF-8",
+//         },
+//     })
+//         .then((response) => response.json())
+//         // .then((json) => console.log(json));
+//         .then((json) => respondReplEntry(json));
+// }
+
+// readLine();